lot of updates
Former-commit-id: d11847f844a3160fa94fe697aef42ca54a16a713
Former-commit-id: 86765259931280cc5cf70c56fb9b96c5f59eabba

+++ charts/ios/Classes/AkylasChartsPlotProxy.m
@@ -176,12 +176,12 @@
 	[self performSelectorOnMainThread:@selector(reconfigurePlot) withObject:nil waitUntilDone:NO];
 }
 
--(void)addData:(NSArray*)values startingAtIndex:(int)index
+-(void)addData:(NSArray*)values startingAtIndex:(NSUInteger)index
 {
 	if (values != nil) {
 		Class dictionaryClass = [NSDictionary class];
 		Class arrayClass = [NSArray class];
-		int length = [values count];
+		NSUInteger length = [values count];
         if (dataX == nil) {
             dataX = [[NSMutableArray arrayWithCapacity:length] retain];
         }
